+++ Assets/Tests/InputSystem/Plugins/PlayerInputTests.cs
@@ -1500,11 +1500,19 @@ internal class PlayerInputTests : InputTestFixture
         var gamepad3 = InputSystem.AddDevice<Gamepad>();
         var gamepad4 = InputSystem.AddDevice<Gamepad>();
 
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true),
+            Is.EquivalentTo(new[] { gamepad1, gamepad2, gamepad3, gamepad4 }));
+
         // Join two players and make sure we get two screen side-by-side.
         Press(gamepad1.buttonSouth);
         Press(gamepad2.buttonSouth);
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(2));
+
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad2 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true),
+            Is.EquivalentTo(new[] { gamepad3, gamepad4 }));
 
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(1));
@@ -1525,6 +1533,12 @@ internal class PlayerInputTests : InputTestFixture
         Press(gamepad3.buttonSouth);
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(3));
+
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad2 }));
+        Assert.That(PlayerInput.all[2].devices, Is.EquivalentTo(new[] { gamepad3 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true),
+            Is.EquivalentTo(new[] { gamepad4 }));
 
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(1));
@@ -1552,6 +1566,12 @@ internal class PlayerInputTests : InputTestFixture
         Press(gamepad4.buttonSouth);
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(4));
+
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad2 }));
+        Assert.That(PlayerInput.all[2].devices, Is.EquivalentTo(new[] { gamepad3 }));
+        Assert.That(PlayerInput.all[3].devices, Is.EquivalentTo(new[] { gamepad4 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true), Is.Empty);
 
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(1));
@@ -1587,6 +1607,12 @@ internal class PlayerInputTests : InputTestFixture
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(3));
 
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad3 }));
+        Assert.That(PlayerInput.all[2].devices, Is.EquivalentTo(new[] { gamepad4 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true),
+            Is.EquivalentTo(new[] { gamepad2 }));
+
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(2));
         Assert.That(PlayerInput.all[2].splitScreenIndex, Is.EqualTo(3));
@@ -1613,6 +1639,13 @@ internal class PlayerInputTests : InputTestFixture
         Press(gamepad2.buttonSouth);
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(4));
+
+        // PlayerInput.all is sorted by playerIndex so the player we just joined popped up in slot #1.
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad2 }));
+        Assert.That(PlayerInput.all[2].devices, Is.EquivalentTo(new[] { gamepad3 }));
+        Assert.That(PlayerInput.all[3].devices, Is.EquivalentTo(new[] { gamepad4 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true), Is.Empty);
 
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(1));
@@ -1648,6 +1681,13 @@ internal class PlayerInputTests : InputTestFixture
         Press(gamepad5.buttonSouth);
 
         Assert.That(PlayerInput.all, Has.Count.EqualTo(5));
+
+        Assert.That(PlayerInput.all[0].devices, Is.EquivalentTo(new[] { gamepad1 }));
+        Assert.That(PlayerInput.all[1].devices, Is.EquivalentTo(new[] { gamepad2 }));
+        Assert.That(PlayerInput.all[2].devices, Is.EquivalentTo(new[] { gamepad3 }));
+        Assert.That(PlayerInput.all[3].devices, Is.EquivalentTo(new[] { gamepad4 }));
+        Assert.That(PlayerInput.all[4].devices, Is.EquivalentTo(new[] { gamepad5 }));
+        Assert.That(InputUser.GetUnpairedInputDevices().ToArray(dispose: true), Is.Empty);
 
         Assert.That(PlayerInput.all[0].splitScreenIndex, Is.EqualTo(0));
         Assert.That(PlayerInput.all[1].splitScreenIndex, Is.EqualTo(1));